+++ libs/buttons/jacdac.ts
@@ -2,7 +2,7 @@ namespace jacdac {
     export class ButtonHostDriver extends SensorHostDriver {
         private button: Button;
         constructor(name: string, button: Button) {
-            super(name, jacdac.BUTTON_DRIVER_CLASS);
+            super(name, jacdac.BUTTON_DEVICE_CLASS);
             this.button = button;
             jacdac.BUTTON_EVENTS.forEach((ev, j) => {
                 control.onEvent(this.button.id(), ev, () => {
