+++ libs/microphone/pxt.json
@@ -4,6 +4,7 @@
     "files": [
         "README.md",
         "microphone.cpp",
+        "jacdac.ts",
         "ns.ts",
         "enums.d.ts",
         "shims.d.ts"
@@ -13,6 +14,7 @@
     ],
     "public": true,
     "dependencies": {
-        "core": "file:../core"
+        "core": "file:../core",
+        "jacdac": "file:../jacdac"
     }
 }
